Arabic dataset preprocessing functions and pipeline

+++ dataset_.py
@@ -79,6 +79,6 @@ class BilingualDataset(Dataset):
         }
 
 def causal_mask(size):
-    mask = torch.triu(torch.ones((1, size, size)), diagonal=1).type(torch.int)
+    mask = torch.triu(torch.ones((1, size, size)), diagonal=1).type(torch.int64)
     return mask == 0
